Removendo usuários sem remover admin-master

+++ app/Http/Livewire/Kuber/DatatablesUser.php
@@ -5,14 +5,24 @@ namespace App\Http\Livewire\Kuber;
 use App\Http\Livewire\Kuber\datatables\ComponentDatatables;
 use App\Models\User;
 use Illuminate\Support\Facades\Auth;
+use Jantinnerezo\LivewireAlert\LivewireAlert;
 use Livewire\Component;
 
 class DatatablesUser extends ComponentDatatables
 {
+    use LivewireAlert;
+    
     public function delete($id)
     {
-        User::where('id', $id)->delete();
+        $user = User::find($id);
 
+        if ($user == null || $user->hasRole('admin-master') == true) {
+            $this->alert('error', 'O super usuário não pode ser removido');
+            return 'error-delete-user';
+        }
+
+        User::where('id', $id)->delete();
+            
         foreach ($this->data as $key => $value) {
             if ($this->data[$key]->id == $id) {
                 $this->data->forget($key);
@@ -21,6 +31,9 @@ class DatatablesUser extends ComponentDatatables
 
         $this->updateData($this->data);
         $this->bootstrap();
+
+        $this->alert('success', 'Usuário removido com sucesso');
+        return 'success-delete-user';
     }
 
     public function render()
